requirements.txt
blinker==1.9.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.4
Flask==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==4.9.4
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pycparser==2.22
PyYAML==6.0.2
requests==2.32.4
stripe==12.2.0
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
xmlsec==1.3.13
gunicorn==23.0.0
boto3==1.35.32

# Development dependencies
black==23.12.1
isort==5.13.2
flake8==6.1.0
mypy==1.8.0
types-requests==2.31.0.20240125
types-PyYAML==6.0.12.20240311
